+++ src/Makefile.am
@@ -304,6 +304,7 @@ libbitcoin_consensus_a_SOURCES = \
   consensus/merkle.h \
   consensus/params.h \
   consensus/validation.h \
+  feerate.h \
   hash.cpp \
   hash.h \
   prevector.h \
@@ -345,6 +346,7 @@ libbitcoin_common_a_SOURCES = \
   coins.cpp \
   compressor.cpp \
   dstencode.cpp \
+  feerate.cpp \
   globals.cpp \
   core_read.cpp \
   core_write.cpp \
